backend/requirements.txt
# Backend dependencies

langchain>=0.2.2
langchainhub
langchain-openai
openai
sentence-transformers
faiss-cpu
python-dotenv
PyPDF2
python-docx
pandas
numpy
fastapi
uvicorn
websockets
python-multipart
langchain
langchain-community
fastapi
uvicorn
python-multipart
openpyxl
python-docx
Pillow
PyPDF2
sentence-transformers
faiss-cpu
pydantic

# Gradio is optional ( to keep the old interface )
gradio>=4.0.0